Updated file tree and renamed routes.py to be login_routes.py
